pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Davi Gabriel Tomaz", email = "davigabtomaz@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]
